break up rcomp
