--- tools/CMakeLists.txt
@@ -1,7 +1,0 @@
-cmake_minimum_required(VERSION 3.12)
-project(tools)
-
-set(CMAKE_CXX_STANDARD 11)
-
-add_executable(tools
-    rcomp.cpp)
